Upload to the CDN before publishing to npm
This simplifies the build process and ensures that new versions of assets will
be available before someone could try to use them through the new version
of the npm module. This change also means that assets will be published to
npm, but that’s not a bad thing, maybe a good thing even.