Wait until the sequence picker iframe logs in before messaging it
If we immediately postMessage the iframe after it loads, but the iframe loads
to the login page rather than to the sequence picker, we’ll drop the message.
Wait until the sequence picker signals us that it’s ready—which also tells us
that the iframe has loaded, so we can replace the old check.